requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# HTTP client for API calls
httpx==0.25.2
aiohttp==3.9.1

# API client libraries
openai==1.3.0
anthropic==0.7.0
google-generativeai==0.3.0

# Basic ML dependencies (lightweight)
numpy==1.24.3
pandas==2.0.3

# JSON handling
jsonlines==4.0.0

# Environment and configuration
python-dotenv==1.0.0

# Note: Heavy dependencies like torch, transformers, etc. are excluded
# as they exceed Vercel's deployment limits. This version focuses on
# API-based inference using external services.
